Project/AL-S-AI/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
blas=1.0=mkl
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.08.22=h06a4308_0
certifi=2021.5.30=py36h06a4308_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.11.0=pyhd3eb1b0_0
dataclasses=0.8=pyh4f3eec9_6
dbus=1.13.18=hb2f20db_0
expat=2.5.0=h6a678d5_0
ffmpeg=4.3=hf484d3e_0
fontconfig=2.14.1=h52c9d5c_1
freetype=2.12.1=h4a9f257_0
glib=2.69.1=h4ff587b_1
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
gst-plugins-base=1.14.1=h6a678d5_1
gstreamer=1.14.1=h5eee18b_1
icu=58.2=he6710b0_3
intel-openmp=2022.1.0=h9e868ea_3769
joblib=1.0.1=pyhd3eb1b0_0
jpeg=9e=h5eee18b_1
kiwisolver=1.3.1=py36h2531618_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libdeflate=1.17=h5eee18b_0
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libuv=1.44.2=h5eee18b_0
libwebp-base=1.3.2=h5eee18b_0
libxcb=1.15=h7f8727e_0
libxml2=2.9.14=h74e7548_0
lz4-c=1.9.4=h6a678d5_0
matplotlib=3.3.4=py36h06a4308_0
matplotlib-base=3.3.4=py36h62a2d02_0
mkl=2020.2=256
mkl-service=2.3.0=py36he8ac12f_0
mkl_fft=1.3.0=py36h54f3939_0
mkl_random=1.1.1=py36h0573a6f_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
numpy=1.19.2=py36h54aff64_0
numpy-base=1.19.2=py36hfa32c7d_0
olefile=0.46=py36_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openssl=1.1.1w=h7f8727e_0
pandas=1.1.5=py36ha9443f7_0
pcre=8.45=h295c915_0
pillow=8.3.1=py36h2c7a002_0
pip=21.2.2=py36h06a4308_0
pyparsing=3.0.4=pyhd3eb1b0_0
pyqt=5.9.2=py36h05f1152_2
python=3.6.13=h12debd9_1
python-dateutil=2.8.2=pyhd3eb1b0_0
pytorch=1.10.2=py3.6_cuda11.3_cudnn8.2.0_0
pytorch-mutex=1.0=cuda
pytz=2021.3=pyhd3eb1b0_0
qt=5.9.7=h5867ecd_1
readline=8.2=h5eee18b_0
scikit-learn=0.24.2=py36ha9443f7_0
scipy=1.5.2=py36h0b6359f_0
setuptools=58.0.4=py36h06a4308_0
sip=4.19.8=py36hf484d3e_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.41.2=h5eee18b_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.12=h1ccaba5_0
torchvision=0.11.3=py36_cu113
tornado=6.1=py36h27cfd23_0
typing_extensions=4.1.1=pyh06a4308_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.4.2=h5eee18b_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
